requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
autogen==0.3.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
et-xmlfile==1.1.0
fastapi==0.115.0
FLAML==2.3.1
frozendict==2.4.4
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
lxml==5.3.0
multidict==6.1.0
multitasking==0.0.11
numpy==1.26.4
openai==1.51.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.17.6
platformdirs==4.3.6
polygon-api-client==1.14.2
propcache==0.2.0
py_lets_be_rational==1.0.1
py_vollib==1.0.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
termcolor==2.4.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
webencodings==0.5.1
websockets==12.0
yarl==1.15.5
yfinance==0.2.44
